Feature/nep (#114)
* Refactor traits to closely match NEPs

* Implement pagination to fetch tokens

* Fix unit tests as they assume n is 64 bits

* Update ts bindings

* Add unit tests for metadata

* Update bindings & docs

* Improve collectible docs

+++ mg-core/src/mocked_context.rs
@@ -107,9 +107,9 @@ pub fn market() -> ValidAccountId {
     "market".try_into().unwrap()
 }
 
-pub fn gate_id(n: u64) -> ValidGateId {
+pub fn gate_id(n: u16) -> ValidGateId {
     let mut hasher = Sha256::new();
-    hasher.update(n.to_ne_bytes());
+    hasher.update((n as u64).to_ne_bytes());
     let result = hasher.finalize();
     let data: &[u8] = result[..16].try_into().unwrap();
     bs58::encode(data).into_string().try_into().unwrap()
